fix(resume-declaration-modal): move resume-declaration-modal to navbar component + links fix
